# controllers/product.controller.js
const catchAsync = require('../utils/catchAsync');
const productService = require('../services/product.service');

const createProduct = catchAsync(async (req, res) => {
  const data = req.body;
  if (req.files) {
    data.images = req.files.map((file) => `/${file.path.replace('\\', '/')}`);
  }
  const product = await productService.createProduct(data);
  res.status(201).json({ status: 'success', data: product });
});

const getProducts = catchAsync(async (req, res) => {
  const result = await productService.queryProducts(req.query);
  res.status(200).json({ status: 'success', ...result });
});

const getProduct = catchAsync(async (req, res) => {
  const product = await productService.getProductById(req.params.id);
  res.status(200).json({ status: 'success', data: product });
});

const updateProduct = catchAsync(async (req, res) => {
  const data = req.body;
  if (req.files) {
    data.images = req.files.map((file) => `/${file.path.replace('\\', '/')}`);
  }
  const product = await productService.updateProduct(req.params.id, data);
  res.status(200).json({ status: 'success', data: product });
});

const deleteProduct = catchAsync(async (req, res) => {
  await productService.deleteProduct(req.params.id);
  res.status(204).json({ status: 'success', data: null });
});

module.exports = { createProduct, getProducts, getProduct, updateProduct, deleteProduct };