# docs/swagger.yaml
openapi: 3.0.3
info:
  title: E-commerce API
  version: 1.0.0
  description: Production-ready backend for an e-commerce application built with Node.js, Express, and MongoDB.
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
paths:
...
  /products:
    get:
      summary: Retrieve all products with pagination, filtering, and sorting
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: sortBy
          schema:
            type: string
      responses:
        '200':
          description: Product list returned
    post:
      summary: Create a new product (admin)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                category:
                  type: string
                stock:
                  type: integer
                images:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Product created
  /products/{id}:
    get:
      summary: Get product details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product found
    put:
      summary: Update product (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                category:
                  type: string
                stock:
                  type: integer
                images:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Product updated
    delete:
      summary: Delete product (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Product deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []